requirements.txt
absl-py==2.3.0
accelerate==1.6.0
asttokens==3.0.0
astunparse==1.6.3
audioread==3.0.1
av==14.3.0
beartype==0.18.5
beautifulsoup4==4.13.4
bitsandbytes==0.45.5
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
CoLT5-attention==0.11.1
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
decorator==4.4.2
decord==0.6.0
einops==0.8.0
einops-exts==0.0.4
einx==0.3.0
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.18.0
fire==0.7.0
Flask==3.1.1
flask-cors==6.0.0
flatbuffers==25.2.10
fonttools==4.57.0
frozendict==2.4.6
fsspec==2025.3.2
fvcore==0.1.5.post20221221
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.72.1
gunicorn==23.0.0
h5py==3.9.0
hf-xet==1.1.0
huggingface-hub==0.31.1
hyper-connections==0.1.15
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_resources==6.5.2
iopath==0.1.10
ipython==8.36.0
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.0
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.44.0
local-attention==1.11.1
loguru==0.7.3
lxml==5.4.0
lz4==4.4.4
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
mtcnn==1.0.0
namex==0.1.0
networkx==3.4.2
nibabel==5.3.2
nilearn==0.9.2
numba==0.61.2
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.8.0.76
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==1.5.1
parameterized==0.9.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
pooch==1.8.2
portalocker==3.1.1
proglog==0.1.11
prompt_toolkit==3.0.51
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytorchvideo==0.1.5
pytz==2025.2
PyYAML==6.0.2
qwen==0.1.1
regex==2024.11.6
requests==2.32.3
retina-face==0.0.17
rich==13.8.0
rotary-embedding-torch==0.3.5
safetensors==0.5.3
scikit-learn==1.1.1
scipy==1.9.3
seaborn==0.13.2
six==1.17.0
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
stack-data==0.6.3
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
threadpoolctl==3.6.0
timm==1.0.15
tokenizers==0.19.1
torch==2.7.1
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
triton==3.3.1
typing_extensions==4.13.2
urllib3==2.4.0
vector-quantize-pytorch==1.17.1
wcwidth==0.2.13
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.2
xformers==0.0.31
yacs==0.1.8
zetascale==2.8.2
